Reduction tests go to their own module
